merge

+++ PersianCalendar/src/main/java/com/byagowi/persiancalendar/reminder/model/Reminder.java
@@ -8,15 +8,14 @@ import java.util.concurrent.TimeUnit;
  */
 public class Reminder {
 
-    final public long id;
+    final public int id;
     final public String name;
     final public String info;
     final public TimeUnit unit;
     final public int quantity;
     final public long startTime;
 
-    public Reminder(long id, String name, String info, TimeUnit unit, int quantity,
-                    long startTime) {
+    public Reminder(int id, String name, String info, TimeUnit unit, int quantity, long startTime) {
         this.id = id;
         this.name = name;
         this.info = info;
